--- eta-unit-testing-project-2019.2-Recife-main/.gitignore
@@ -1,16 +1,0 @@
-# Eclipse
-.classpath
-.project
-.settings/
-
-# Intellij
-.idea/
-*.iml
-*.iws
-
-# Mac
-.DS_Store
-
-# Maven
-log/
-target/
