--- eta-unit-testing-project-2019.2-Recife-main/README.md
@@ -1,12 +1,0 @@
-# [ETA] Unit Testing Project<br> 2019.2-Recife
-
-[![Java CI with Maven](https://github.com/cesar-school/eta-unit-testing-project-2019.2-Recife/workflows/CI/badge.svg)](https://github.com/cesar-school/eta-unit-testing-project-2019.2-Recife/actions?query=workflow%3A%22CI%22)
-[![Coverage](https://sonarcloud.io/api/project_badges/measure?project=cesar-school_eta-unit-testing-project-2019.2-Recife&metric=coverage)](https://sonarcloud.io/dashboard?id=cesar-school_eta-unit-testing-project-2019.2-Recife)
-
-O projeto contém 11 testes, que valem 1 ponto cada se implementado de forma correta. O aluno que escrever corretamente 10 dos 11 testes fará jus a nota máxima.
-
-Usando o Github, cada aluno deverá fazer um “fork” do repositório do projeto.
-
-No arquivo `src\test\java\school\cesar\eta\unit\PersonTest.java` estão as assinaturas dos métodos de teste, seguindo o padrão de nomenclatura visto em aula ` UnitOfWork_StateUnderTest_ExpectedBehavior`. Cada aluno terá 15 dias e, de forma individual, deverá escrever o corpo dos métodos de teste relacionados a classe `Person`, que fica em `src\main\java\school\cesar\eta\unit\Person.java`. Nenhuma modificação deverá ser feita em `Person`, a classe testada.
-
-Uma vez concluído a escrita dos testes, o aluno deverá fazer um “commit” em seu repositório e abrir um “pull request” tendo o repositório original como destino. Será por meio dele que o aluno será avaliado e o feedback será dado.
