+++ src/test/java/school/cesar/eta/unit/FakePerson.java
@@ -1,3 +1,4 @@
+package school.cesar.eta.unit;
 
 import school.cesar.eta.unit.Person;
